Proposed changes to move wifi func to a single header solution
