+++ src/main.cpp
@@ -1,7 +1,6 @@
 #include <Arduino.h>
-#include <WiFiS3.h>
-
-#include "arduino_secrets.h"
+#include "wifi_arduino.h"
+#include "clap_detection_example.h"
 
 #define PhotoResistor_PIN 2                 // Will eventually be digital input pin for light-sensor
 #define SoundAnalog_PIN A1                  // Will eventually be analog input pin for sound-sensor
@@ -9,16 +8,8 @@
 #define LED_PIN 4                           // Eventual digital-output pin a light-source
 #define Measured_Light_Value 0              // Eventually analog-input from light sensor
 #define Sound_Treshold 500                  // Will serve as calibration for our sound-sensor
-#define WIFI_RECONNECTION_ATTEMPTS 10       // Used in connectToWifi()
-#define WIFI_TIME_BETWEEN_RECONNECTION 1000 // Defined in milliseconds. Used in connectToWifi()
 
 bool isLampOn = false;
-
-// YOU NEED THIS IN A SEPERATE "arduino_secrets.h" FILE OR IT WILL NOT WORK/COMPILE
-//
-// -----  WiFi Settings  ----
-// const char SSID[] = "SSID_NAME_HERE";           // your WiFi network name
-// const char PASSWORD[] = "WIFI_PASSWORD_HERE";   // your WiFi password
 
 // Function Declarations:
 
@@ -27,86 +18,47 @@ void Sound_toggleLampOnClap();
 
 void setup()
 {
-  Serial.begin(115200);
-  pinMode(LED_PIN, OUTPUT);
-  pinMode(PhotoResistor_PIN, INPUT);
-  pinMode(SoundAnalog_PIN, INPUT);
-  pinMode(SoundDigital_PIN, INPUT);
+    Serial.begin(115200);
+    pinMode(LED_PIN, OUTPUT);
+    pinMode(PhotoResistor_PIN, INPUT);
+    pinMode(SoundAnalog_PIN, INPUT);
+    pinMode(SoundDigital_PIN, INPUT);
 }
 
 void loop()
 {
-  Serial.print("Light Intensity: ");
-  Serial.println(digitalRead(PhotoResistor_PIN));
-  Serial.print("Sound Intensity: ");
-  Serial.println(analogRead(SoundAnalog_PIN));
+    Serial.print("Light Intensity: ");
+    Serial.println(digitalRead(PhotoResistor_PIN));
+    Serial.print("Sound Intensity: ");
+    Serial.println(analogRead(SoundAnalog_PIN));
 
-  Sound_toggleLampOnClap();
+    Sound_toggleLampOnClap();
+    detect_claps(SoundAnalog_PIN, LED_PIN);
 
-  // digitalWrite(LED_PIN, HIGH);
-  // if (digitalRead(PhotoResistor_PIN) == HIGH)
-  // {
-  //   digitalWrite(LED_PIN, HIGH);
-  // }
-  // else
-  // {
-  //   digitalWrite(LED_PIN, LOW);
-  // }
+    // digitalWrite(LED_PIN, HIGH);
+    // if (digitalRead(PhotoResistor_PIN) == HIGH)
+    // {
+    //   digitalWrite(LED_PIN, HIGH);
+    // }
+    // else
+    // {
+    //   digitalWrite(LED_PIN, LOW);
+    // }
 
-  delay(100);
+    delay(100);
 }
 
 // Function Definitions:
 
 void Sound_toggleLampOnClap()
 {
-  if (analogRead(SoundAnalog_PIN) < Sound_Treshold)
-  {
-    isLampOn = !isLampOn;
-    digitalWrite(LED_PIN, isLampOn);
-  }
-  else
-  {
-    return;
-  }
-}
-
-/**
- * @brief Meant to be used in setup() and does NOT handle errors
- */
-void ConnectToWifi()
-{
-  WiFi.disconnect(); // Ensures a clean start
-  WiFi.end();        // Hard reset of the WiFi module
-  delay(1000);       // Give it time(ms) to reset
-
-  int numberOfAttempts = 0;
-
-  while (WiFi.status() != WL_CONNECTED)
-  {
-    Serial.println("Attempting to connect to Wi-Fi.");
-    WiFi.begin(SSID, PASSWORD);
-    if (WiFi.status() == WL_CONNECTED)
+    if (analogRead(SoundAnalog_PIN) < Sound_Treshold)
     {
-      Serial.println("Connected!");
-      numberOfAttempts = 0;
-      return;
+        isLampOn = !isLampOn;
+        digitalWrite(LED_PIN, isLampOn);
     }
-    Serial.print("Attempt: ");
-    Serial.print(numberOfAttempts + 1);
-    Serial.print("\n");
-
-    Serial.print("Wi-Fi Status Code: ");
-    Serial.print(WiFi.status());
-    Serial.print("\n");
-
-    if (numberOfAttempts >= WIFI_RECONNECTION_ATTEMPTS)
+    else
     {
-      Serial.println("Connection failed.");
-      return;
+        return;
     }
-
-    numberOfAttempts++;
-    delay(WIFI_TIME_BETWEEN_RECONNECTION);
-  }
 }
